+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Improvement: Support for monotonic clock on Android.
   [P-R #5](https://github.com/bbc/pydvbcss/pull/5) by [Jack Jansen](https://github.com/jackjansen) (03 May 2016)
+* Bugfix: `clock.getEffectiveSpeed()` stuck in infinite loop.
 
 ## 0.3.3 : Bugfixes and thread safety improvements (20 Apr 2016)